segregate MethodHandles code from aggregation logic to make it easier to delete when increasing the language level
